feat: detect if Pinpoint project exists before prompt for remove command
feat: detect if Pinpoint project exists before prompt for remove command

+++ packages/amplify-category-notifications/commands/notifications/remove.js
@@ -11,51 +11,56 @@ module.exports = {
   alias: ['disable', 'delete'],
   run: async (context) => {
     context.exeInfo = context.amplify.getProjectDetails();
-    const availableChannels = notificationManager.getAvailableChannels(context);
-    const enabledChannels = notificationManager.getEnabledChannels(context);
+    const pinpointApp = pinpointHelper.getPinpointApp(context);
+    if (pinpointApp) {
+      const availableChannels = notificationManager.getAvailableChannels(context);
+      const enabledChannels = notificationManager.getEnabledChannels(context);
 
-    enabledChannels.push(PinpointApp);
-    enabledChannels.push(Cancel);
+      enabledChannels.push(PinpointApp);
+      enabledChannels.push(Cancel);
 
-    let channelName = context.parameters.first;
+      let channelName = context.parameters.first;
 
-    if (!channelName || !availableChannels.includes(channelName)) {
-      const answer = await inquirer.prompt({
-        name: 'selection',
-        type: 'list',
-        message: 'Choose what to remove.',
-        choices: enabledChannels,
-        default: enabledChannels[0],
-      });
-      channelName = answer.selection;
-    } else if (!enabledChannels.includes(channelName)) {
-      context.print.info(`The ${channelName} channel has NOT been enabled.`);
-      channelName = undefined;
-    }
-
-    if (channelName && channelName !== Cancel) {
-      if (channelName !== PinpointApp) {
-        await pinpointHelper.ensurePinpointApp(context);
-        await notificationManager.disableChannel(context, channelName);
-        writeAmplifyMeta(context);
-      } else if (pinpointHelper.isAnalyticsAdded(context)) {
-        context.print.error('Execution aborted.');
-        context.print.info('You have an analytics resource in your backend tied to the Amazon Pinpoint resource');
-        context.print.info('The Analytics resource must be removed before Amazon Pinpoint can be deleted from the cloud');
-      } else {
+      if (!channelName || !availableChannels.includes(channelName)) {
         const answer = await inquirer.prompt({
-          name: 'deletePinpointApp',
-          type: 'confirm',
-          message: 'Confirm that you want to delete the associated Amazon Pinpoint application',
-          default: false,
+          name: 'selection',
+          type: 'list',
+          message: 'Choose what to remove.',
+          choices: enabledChannels,
+          default: enabledChannels[0],
         });
-        if (answer.deletePinpointApp) {
-          await pinpointHelper.deletePinpointApp(context);
-          context.print.info('The Pinpoint application has been successfully deleted.');
+        channelName = answer.selection;
+      } else if (!enabledChannels.includes(channelName)) {
+        context.print.info(`The ${channelName} channel has NOT been enabled.`);
+        channelName = undefined;
+      }
+      if (channelName && channelName !== Cancel) {
+        if (channelName !== PinpointApp) {
+          await pinpointHelper.ensurePinpointApp(context);
+          await notificationManager.disableChannel(context, channelName);
           writeAmplifyMeta(context);
+        } else if (pinpointHelper.isAnalyticsAdded(context)) {
+          context.print.error('Execution aborted.');
+          context.print.info('You have an analytics resource in your backend tied to the Amazon Pinpoint resource');
+          context.print.info('The Analytics resource must be removed before Amazon Pinpoint can be deleted from the cloud');
+        } else {
+          const answer = await inquirer.prompt({
+            name: 'deletePinpointApp',
+            type: 'confirm',
+            message: 'Confirm that you want to delete the associated Amazon Pinpoint application',
+            default: false,
+          });
+          if (answer.deletePinpointApp) {
+            await pinpointHelper.deletePinpointApp(context);
+            context.print.info('The Pinpoint application has been successfully deleted.');
+            writeAmplifyMeta(context);
+          }
         }
       }
+    } else {
+      context.print.error('Notifications have not been added to your project.');
     }
     return context;
   },
 };
+
